refactor(cli): deduplicate getMergedOptions, getJsonFlag, maybeQuiet
Move getMergedOptions (6 copies), getJsonFlag, withCapturedConsole, and
maybeQuiet into jsonHelpers.ts as the single shared location. Remove the
process.argv fallback from getJsonFlag — Commander option propagation
handles this correctly. Extract expectJsonHelpOption test helper to
reduce boilerplate across --json unit tests.

Addresses review items 3-5, 8 from PR #76.
