+++ packages/cli/src/cli/commands/content.ts
@@ -1,12 +1,16 @@
 import { Command } from "commander";
 import chalk from "chalk";
-import type { CommandOptions } from "../../types/types";
 import { viewDomainContentHash, setDomainContentHash } from "../../commands/contentHash";
 import { addAuthOptions } from "./authOptions";
 import { prepareContext } from "../context";
-import { prepareReadOnlyContext, getJsonFlag } from "./lookup";
-import { maybeQuiet } from "./bulletin";
-import { emitJsonResult, handleCommandError } from "./jsonHelpers";
+import { prepareReadOnlyContext } from "./lookup";
+import {
+  getMergedOptions,
+  getJsonFlag,
+  maybeQuiet,
+  emitJsonResult,
+  handleCommandError,
+} from "./jsonHelpers";
 import ora from "ora";
 
 export interface ContentViewOptions {
@@ -15,25 +19,6 @@ export interface ContentViewOptions {
 
 export interface ContentSetOptions {
   rpc?: string;
-}
-
-function getMergedOptions<T>(command: Command | undefined, fallback: T): CommandOptions & T {
-  const mergedOptions: any = { ...(fallback ?? {}) };
-
-  let currentCommand: Command | null | undefined = command?.parent;
-  while (currentCommand) {
-    if (typeof currentCommand.opts === "function") {
-      const parentOptions = currentCommand.opts();
-      for (const key in parentOptions) {
-        if (!(key in mergedOptions) && parentOptions[key] !== undefined) {
-          mergedOptions[key] = parentOptions[key];
-        }
-      }
-    }
-    currentCommand = currentCommand.parent;
-  }
-
-  return mergedOptions;
 }
 
 export function attachContentCommands(root: Command) {
@@ -56,9 +41,6 @@ export function attachContentCommands(root: Command) {
         );
 
         if (!jsonOutput) console.log(chalk.bold("\n▶ Content View\n"));
-        // Ora caches process.stderr (the object), not .write (the method).
-        // withCapturedConsole replaces .write on the object, so spinner
-        // output is captured as long as start/succeed/fail run inside maybeQuiet.
         const spinner = ora();
 
         const result = await maybeQuiet(jsonOutput, () =>
@@ -95,7 +77,6 @@ export function attachContentCommands(root: Command) {
         );
 
         if (!jsonOutput) console.log(chalk.bold("\n▶ Content Set\n"));
-        // See view handler above for why ora is safe inside maybeQuiet.
         const spinner = ora();
 
         const result = await maybeQuiet(jsonOutput, () =>
