+++ packages/cli/tests/unit/register/registerJson.test.ts
@@ -1,25 +1,15 @@
 import { expect, test } from "bun:test";
 import {
-  HARNESS_HELP_SUCCESS_EXIT_CODE,
   HARNESS_SUCCESS_EXIT_CODE,
   runDotnsCli,
+  expectJsonHelpOption,
 } from "../../_helpers/cliHelpers";
 
-test("register domain --help shows --json option", async () => {
-  const result = await runDotnsCli(["register", "domain", "--help"]);
+test("register domain --help shows --json option", () =>
+  expectJsonHelpOption(["register", "domain"]));
 
-  expect(result.exitCode).toBe(HARNESS_HELP_SUCCESS_EXIT_CODE);
-  expect(result.combinedOutput).toContain("--json");
-  expect(result.combinedOutput).toContain("Output result as JSON");
-});
-
-test("register subname --help shows --json option", async () => {
-  const result = await runDotnsCli(["register", "subname", "--help"]);
-
-  expect(result.exitCode).toBe(HARNESS_HELP_SUCCESS_EXIT_CODE);
-  expect(result.combinedOutput).toContain("--json");
-  expect(result.combinedOutput).toContain("Output result as JSON");
-});
+test("register subname --help shows --json option", () =>
+  expectJsonHelpOption(["register", "subname"]));
 
 test("register domain --json emits JSON error when --transfer without --to", async () => {
   const result = await runDotnsCli([
